Initial Commit
This is the initial repo creation which is essential the same one off
script I was using on my raspberry pi 2 to collect data from a couple
Dallas 1 wire temperature sensors. I just turned it into a git repo
and added the other associated bits to make it a real project.