Change default base_topic to dallasMQTT
The default base_topic was previously 'pi_temp_server' which was an
artifact of the early script that dallasMQTT grew out of. But there is
no guarantee that dallasMQTT is running on a raspberry pi or that it is
simply running as a dedicated temperature server. So this commit changes
the default base_topic value to be a bit more general purpose and sets
it to 'dallasMQTT'.

This is a backwards incompatible change, but since the tool is still
young and completely undocumented hopefully there aren't any users out
there relying on this oddly specific default value.